requirements/docs.txt

matplotlib
numpydoc
sphinx
sphinx-gallery
sphinx_rtd_theme
